Readapted runtime to new update ansible message from UI

+++ runtime/runtime.py
@@ -240,30 +240,32 @@ def msg_handling(msg):
         robot_status = 0
     elif msg_type == 'update':
         #initiate necessary shutdown procedures
-        student_proc.terminate()
-        console_proc.terminate()
-        stop_motor()
-        robot_status = 0
+        if robot_status:
+            student_proc.terminate()
+            console_proc.terminate()
+            stop_motor()
 
         #remove everything that is like tar currently in the update folder
-        os.system('rm -f ~/updates/*tar*')
+        #os.system('rm -f ~/updates/*tar*')
 
-        update = b64decode(msg['content']['update'])
-        signature = b64decode(msg['content']['signature'])
-        filename = msg['content']['filename']
-        signature_filename = filename + '.asc'
-        update_f = open(filename, 'wb')
-        update_f.write(bytearray(update))
-        update_f.flush()
-        update_f.close()
-        signature_f = open(signature_filename, 'wb')
-        signature_f.write(bytearray(signature))
-        signature_f.flush()
-        signature_f.close()
+        #update = b64decode(msg['content']['update_path'])
+        #signature = b64decode(msg['content']['signature_path'])
+        #update_path = msg['content']['update_path']
+        #signature_path = msg['content']['signature_path']
+        #filename = msg['content']['filename']
+        #signature_filename = filename + '.asc'
+        #update_f = open(, 'wb')
+        #update_f.write(bytearray(update))
+        #update_f.flush()
+        #update_f.close()
+        #signature_f = open(signature_filename, 'wb')
+        #signature_f.write(bytearray(signature))
+        #signature_f.flush()
+        #signature_f.close()
 
         #put signature_f and update_f in ~/updates
-        os.system('mv ' + filename + ' ~/updates')
-        os.system('mv ' + signature_filename + ' ~/updates')
+        #os.system('mv ' + filename + ' ~/updates')
+        #os.system('mv ' + signature_filename + ' ~/updates')
 
         #run ./update.sh
         os.system('sh ~/updates/update.sh')
